server/requirements.txt
fastapi
uvicorn
python-dotenv
redis
asyncpg
sqlalchemy
httpx